requirements.txt
appdirs==1.4.3
asgiref==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
more-itertools==8.3.0
numpy==1.18.4
packaging==20.4
pandas==1.0.3
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.10
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
whitenoise==5.1.0
